+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 # 0.2.0
 
+* added `isBefore` and `isAfter` method on `Intersect`
 * strong mode compliant
 * added `intersect` and `enclose` method on `Intersect` and `intersectAll` constructor